+++ internal/commands/schema_types_test.go
@@ -1,6 +1,7 @@
 package commands
 
 import (
+	"encoding/json"
 	"io"
 	"net/http"
 	"strings"